Role Playing Game
It is kind of a Number game.
Where the player can go to store, go to cave or fight dragon.
from the go to store player can buy health and weapon.
from the go to cave player can fight slime or fight faged beast.
from the fight dragon player can attack or dodge or run.